Implementation delay computation + add comment
