+++ examples/pdu_exchange.rs
@@ -16,14 +16,14 @@ async fn main() -> std::io::Result<()> {
             master.send();
     })};
     std::thread::sleep(Duration::from_millis(500));
-    
+
     let reg = Field::<u16>::simple(0x1234);
     let slave = 0;
-            
+
     // test read/write
     let received = master.aprd(slave, reg).await;
     master.apwr(slave, reg, received.value).await;
-    
+
     // test simultaneous read/write
     futures::join!(
         async {
@@ -37,7 +37,6 @@ async fn main() -> std::io::Result<()> {
             master.apwr(slave, reg, received.value).await;
         },
     );
-    
+
     Ok(())
 }
-
